Add global actor registration; implement aliases for import names; add details section to handler messages; change actor ID from integer to long

+++ core-local/src/main/java/org/torqlang/core/local/ActorBuilder.java
@@ -17,7 +17,7 @@ import org.torqlang.core.util.SourceSpan;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 /*
  * Note that as we transition forward in the process, we gain properties and loose methods.
@@ -70,7 +70,7 @@ public final class ActorBuilder implements ActorBuilderInit, ActorBuilderReady, 
     private static final Str CFG = Str.of("cfg");
     private static final int TIME_SLICE_1000 = 10_000;
 
-    private static final AtomicInteger nextActorId = new AtomicInteger(0);
+    private static final AtomicLong nextActorId = new AtomicLong(0);
 
     private State state;
